# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tinday API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Tinday Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase-issued JWT access token. Obtained from /api/auth/login or /api/auth/register. Include as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "message": {
            "type": "string",
            "description": "Additional error context"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            },
            "description": "Field-level validation errors (422 only)"
          }
        }
      },
...
      "PublicProfile": {
        "type": "object",
        "description": "Public-facing profile (excludes internal fields like preferences and embeddings)",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "about": {
            "type": "string",
            "nullable": true
          },
          "experience_years": {
            "type": "integer"
          },
          "skills": {
            "type": "array",
            "items": { "type": "string" }
          },
          "roles": {
            "type": "array",
            "items": { "type": "string" }
          },
          "projects": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "match_id": {
            "type": "string",
            "format": "uuid"
          },
          "sender_id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string",
            "description": "Message text content"
          },
          "read_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Timestamp when the message was read by the recipient. null if unread."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "description": "Message content (cannot be empty)"
          }
        }
      },
      "InboxEntry": {
        "type": "object",
        "properties": {
          "match_id": {
            "type": "string",
            "format": "uuid"
          },
          "other_user": {
            "$ref": "#/components/schemas/PublicProfile"
          },
          "latest_message": {
            "type": "object",
            "nullable": true,
            "description": "Most recent message in this conversation, or null if no messages yet",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "sender_id": { "type": "string", "format": "uuid" },
              "content": { "type": "string" },
              "read_at": { "type": "string", "format": "date-time", "nullable": true },
              "created_at": { "type": "string", "format": "date-time" }
            }
          },
          "unread_count": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of unread messages from the other user"
          },
          "last_activity": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the latest message or match creation. Used for sorting."
          },
          "muted": {
            "type": "boolean",
            "description": "Whether the current user has muted notifications for this conversation"
          }
        }
      },
      "MatchUserState": {
        "type": "object",
        "description": "The current user's per-conversation inbox state.",
        "properties": {
          "match_id": { "type": "string", "format": "uuid" },
          "muted": {
            "type": "boolean",
            "description": "Whether notifications for this conversation are muted"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the conversation was soft-hidden, or null if visible. The chat reappears once newer activity arrives."
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - Missing, invalid, or expired JWT token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Missing or invalid authorization header" }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Authenticated user does not have access to this resource",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Access denied: Not a member of this match" }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - The requested resource does not exist",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Not Found", "message": "The requested resource was not found." }
          }
        }
      },
...
      "ValidationError": {
        "description": "Unprocessable Entity - Request validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "error": "Validation error",
              "details": [
                { "field": "email", "message": "Invalid email" },
                { "field": "password", "message": "String must contain at least 8 character(s)" }
              ]
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Internal Server Error" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/messaging/inbox": {
      "get": {
        "tags": ["Messaging"],
        "summary": "Get message inbox",
        "description": "Returns all match conversations for the current user, sorted by most recent activity. Each entry includes:\n- The other user's profile\n- The latest message (or null if no messages exist yet)\n- Count of unread messages from the other user\n- Timestamp of last activity\n\nResults are sorted by `last_activity` descending (most recent first).",
        "operationId": "getInbox",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Inbox with conversations sorted by most recent activity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/InboxEntry" }
                },
                "example": [
                  {
                    "match_id": "d4e5f6a7-b8c9-0123-defa-123456789012",
                    "other_user": {
                      "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                      "name": "John Smith",
                      "avatar_url": "https://example.com/avatar.jpg",
                      "about": "Backend developer",
                      "experience_years": 5,
                      "skills": ["JavaScript", "Node.js"],
                      "roles": ["Backend Developer"],
                      "projects": []
                    },
                    "latest_message": {
                      "id": "e5f6a7b8-c9d0-1234-efab-123456789012",
                      "sender_id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                      "content": "Hey! Would love to connect about your AI projects.",
                      "read_at": null,
                      "created_at": "2026-06-05T13:30:00.000Z"
                    },
                    "unread_count": 2,
                    "last_activity": "2026-06-05T13:30:00.000Z"
                  }
                ]
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/messaging/matches/{matchId}/messages": {
      "get": {
        "tags": ["Messaging"],
        "summary": "Get messages for a match",
        "description": "Returns paginated messages for a specific match, sorted by creation time ascending (oldest first). The authenticated user must be a member of the match (403 if not, 404 if match doesn't exist).\n\n**Auto-read**: Messages sent by the other user that are unread will be marked as read (fire-and-forget update).\n\n**Pagination**: Use `from` (inclusive) and `to` (inclusive) for cursor-based pagination. Default range is 0-49.",
        "operationId": "getMessages",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "description": "UUID of the match",
            "schema": { "type": "string", "format": "uuid" },
            "example": "d4e5f6a7-b8c9-0123-defa-123456789012"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "Starting index for pagination (inclusive, 0-based)",
            "schema": { "type": "integer", "minimum": 0, "default": 0 },
            "example": 0
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "Ending index for pagination (inclusive, 0-based)",
            "schema": { "type": "integer", "minimum": 0, "default": 49 },
            "example": 49
          }
        ],
        "responses": {
          "200": {
            "description": "Messages for the match",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Message" }
                },
                "example": [
                  {
                    "id": "e5f6a7b8-c9d0-1234-efab-123456789012",
                    "match_id": "d4e5f6a7-b8c9-0123-defa-123456789012",
                    "sender_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "content": "Hi John, I saw we matched!",
                    "read_at": "2026-06-05T13:31:00.000Z",
                    "created_at": "2026-06-05T13:00:00.000Z"
                  },
                  {
                    "id": "f6a7b8c9-d0e1-2345-fabc-123456789012",
                    "match_id": "d4e5f6a7-b8c9-0123-defa-123456789012",
                    "sender_id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                    "content": "Hey! Would love to connect about your AI projects.",
                    "read_at": null,
                    "created_at": "2026-06-05T13:30:00.000Z"
                  }
                ]
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Messaging"],
        "summary": "Send a message in a match",
        "description": "Sends a text message in a specific match conversation. The authenticated user must be a member of the match. The message is broadcast to the match room via Supabase Realtime so the recipient receives it instantly if subscribed.\n\n**Real-time delivery**: The recipient should subscribe to the `messages` table via Supabase Realtime for the specific `match_id` to receive messages in real time. The HTTP response returns the created message object.",
        "operationId": "sendMessage",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "description": "UUID of the match to send a message to",
            "schema": { "type": "string", "format": "uuid" },
            "example": "d4e5f6a7-b8c9-0123-defa-123456789012"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendMessageRequest" },
              "examples": {
                "basic": {
                  "summary": "Simple message",
                  "value": {
                    "content": "Hey! Would love to connect about your AI projects."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" },
                "example": {
                  "id": "f6a7b8c9-d0e1-2345-fabc-123456789012",
                  "match_id": "d4e5f6a7-b8c9-0123-defa-123456789012",
                  "sender_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "content": "Hey! Would love to connect about your AI projects.",
                  "read_at": null,
                  "created_at": "2026-06-05T13:30:00.000Z"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/messaging/matches/{matchId}/hidden": {
      "put": {
        "tags": ["Messaging"],
        "summary": "Hide (soft-delete) or restore a conversation",
        "description": "Sets the authenticated user's soft-delete state for a match. Hiding (`hidden: true`) removes the conversation from the user's inbox without deleting the match or any messages; it reappears automatically once the other user sends a new message. Restoring (`hidden: false`) un-hides it immediately, supporting an Undo. Per-user and synced across devices. The user must be a member of the match (403 if not, 404 if it doesn't exist).",
        "operationId": "setMatchHidden",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "description": "UUID of the match",
            "schema": { "type": "string", "format": "uuid" },
            "example": "d4e5f6a7-b8c9-0123-defa-123456789012"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["hidden"],
                "properties": {
                  "hidden": {
                    "type": "boolean",
                    "description": "true to hide (delete) the chat, false to restore it"
                  }
                }
              },
              "examples": {
                "delete": { "summary": "Delete chat", "value": { "hidden": true } },
                "restore": { "summary": "Undo / restore", "value": { "hidden": false } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated conversation state",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MatchUserState" },
                "example": {
                  "match_id": "d4e5f6a7-b8c9-0123-defa-123456789012",
                  "muted": false,
                  "deleted_at": "2026-06-11T10:00:00.000Z"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  }
}